requirements.txt
Flask==3.0.2
PyJWT==2.8.0
pandas==2.2.2
openpyxl==3.1.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
